Create comments app and move comment models from utils
The existing Comment and CommentPost models are moved from the utils app to a
new dedicated comments app for better organization and separation of concerns.
Related migrations and references are updated accordingly.
